added login function
The login function uses RCurl to log into the Achievement Series
website.  If given no arguments, it prompts the user for the username,
password, and site ID.  It also requires an SSL certificate.  The
location of the certificate file can be passed as an argument.  If it is
not, the functions downloads a certificate file and saves it to the
working directory.

The function returns a curl handle, which must be used by other
functions that seek to pull http requests from inside the Achievement
Series site.